requirements.txt
Flask
python-whois
Flask-Cors
